requirement.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.48.0
timm>=0.9.8
numpy>=1.27.0
pandas>=2.1.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.9.0
seaborn>=0.13.2
opencv-python>=4.9.0
tqdm>=4.66.0
requests>=2.31.0
pycocotools>=2.0.10
nltk>=3.9.0
streamlit
transformers
torch
diffusers
accelerate
safetensors
pillow
tqdm
open-clip-torch
whisper
tesserocr
pytesseract
faiss-cpu
modal-client
gTTS
